+++ CHANGELOG.md
@@ -37,6 +37,8 @@ For instructions on upgrading to newer versions, visit
 
 * \#4091 Use sublcass context when calling a scope defined in a superclass. (Edgars Beigarts)
 
+* \#4052 Fixed uniqueness validation on localized fields with no value.
+
 * \#4005 Fixed inclusion of mongoid with Rails components that don't have the Rails environment.
 
 * \#3993 Fixes issue where `dup`/`clone` fails for embedded documents that use store_as without using Mongoid::Atributes::Dynamic